CrashReporter: major changes (remote=>local)
We now open a simple dialog when Natron crashes. We will parse the crash dump using symbols included with Natron (../Resources/symbols). Then generate and open a HTML document with the crash report.

This will bloat the install size with ~20MB since we need to distribute the symbols, but I think it's worth it since we don't need to have a crash report server.

The CrashReporter now depends on breakdown (https://github.com/rodlie/breakdown) to be able to parse the crash dump. This can be removed in the future when/if Natron moves from http://github.com/NatronGitHub/google-breakpad to https://github.com/NatronGitHub/breakpad, or Crashpad.